pyproject.toml
[project]
name = "github-issue-management"
version = "0.1.0"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[azure-ai-inference,tools]==1.4.1",
    "PyGithub>=2.1.1",
]

[project.scripts]
run_crew = "github_issue_management.main:run"
train = "github_issue_management.main:train"
replay = "github_issue_management.main:replay"
test = "github_issue_management.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

